requirements.txt
Django==1.10.1
Pygments==2.1.3
djangorestframework==3.4.6